requirements.txt
requests
beautifulsoup4
lxml
pdfminer.six